starting to make parallelization possible

+++ EM_Creator/ma5Wrapper.py
@@ -8,7 +8,8 @@
 .. moduleauthor:: Wolfgang Waltenberger <wolfgang.waltenberger@gmail.com>
 """
 
-import os, sys, colorama, subprocess, shutil
+import os, sys, colorama, subprocess, shutil, tempfile
+import bakeryHelpers
 
 class MA5Wrapper:
     def __init__ ( self, ver="1.7" ):
@@ -24,7 +25,8 @@ class MA5Wrapper:
             self.info ( "cannot find ma5 installation at %s" % self.ma5install )
             self.exe ( "ma5/make.py" )
         self.templateDir = "templates/"
-        self.commandfile = "ma5commands.txt"
+        # self.commandfile = "ma5commands.txt"
+        self.commandfile = tempfile.mktemp ( prefix="ma5", dir="./" )
         self.info ( "initialised" )
 
     def info ( self, *msg ):
@@ -56,7 +58,7 @@ class MA5Wrapper:
         shutil.copy ( templatefile, filename )
         self.info ( "wrote recasting card %s" % filename )
 
-    def writeCommandFile ( self, hepmcfile ):
+    def writeCommandFile ( self, hepmcfile, process, masses ):
         """ this method writes the commands file for ma5.
         :param hepmcfile: I think thats the input events
         """
@@ -64,21 +66,30 @@ class MA5Wrapper:
         f.write('set main.recast = on\n')
         f.write('set main.recast.card_path = recasting_card.dat\n' )
         f.write('import '+hepmcfile+'\n')
-        f.write('submit'+'\n')
+        f.write('submit ANA_%s\n' % bakeryHelpers.dirName(process,masses)  )
         f.close()
 
-    def run( self, hepmcfile, topo ):
+    # def run( self, hepmcfile, topo ):
+    def run( self, process, masses, topo ):
         """ Run MA5 over an hepmcfile, specifying the process """
         self.writeRecastingCard ()
         # then write command file
+        Dir = bakeryHelpers.dirName ( process, masses ) 
+        hepmcfile = "%s/Events/run_01/tag_1_pythia8_events.hepmc.gz" % Dir
         hepmcfile = os.path.abspath ( hepmcfile )
-        self.writeCommandFile( hepmcfile )
+        print ( "hepmcfile at", hepmcfile )
+        if not os.path.exists ( hepmcfile ):
+            print ( "Error cannot find hepmc file at %s" % hepmcfile )
+            sys.exit()
+        self.writeCommandFile( hepmcfile, process, masses )
         # then run madgraph5
         os.chdir ( "ma5/" )
         cmd = "%s -R -s %s 2>&1 | tee /tmp/ma5.run" % (self.executable, \
                 self.commandfile )
         self.exe ( cmd )
-        shutil.move ( "ANALYSIS_0", "ANA_%s" % topo )
+        if os.path.exists ( self.commandfile ):
+            subprocess.getoutput ( "rm -r %s" % self.commandfile )
+        #shutil.move ( "ANALYSIS_0", "ANA_%s" % Dir )
         os.chdir ( "../" )
 
     def exe ( self, cmd ):
@@ -95,4 +106,8 @@ class MA5Wrapper:
 
 if __name__ == "__main__":
     ma5 = MA5Wrapper()
-    ma5.run( "T2tt_1jet/Events/run_01/tag_1_pythia8_events.hepmc.gz", "T2tt" )
+    process = "T2tt_1jet"
+    topo = "T2tt"
+    masses= [500, 100]
+    ma5.run( process, masses, topo )
+    # ma5.run( "T2tt_1jet/Events/run_01/tag_1_pythia8_events.hepmc.gz", "T2tt" )
